+++ plotoptix/tests/test_090_buffers_readback.py
@@ -67,7 +67,7 @@ class TestCallbacks(TestCase):
 
         TestCallbacks.scene.start()
         self.assertTrue(TestCallbacks.scene.is_started(), msg="Scene did not flip to _is_started=True state.")
-        self.assertTrue(TestCallbacks.scene.isAlive(), msg="Raytracing thread is not alive.")
+        self.assertTrue(TestCallbacks.scene.is_alive(), msg="Raytracing thread is not alive.")
         TestCallbacks.is_alive = True
 
         t = 0;
@@ -87,7 +87,7 @@ class TestCallbacks(TestCase):
         TestCallbacks.scene.close()
         TestCallbacks.scene.join(10)
         self.assertTrue(TestCallbacks.scene.is_closed(), msg="Scene did not flip to _is_closed=True state.")
-        self.assertFalse(TestCallbacks.scene.isAlive(), msg="Raytracing thread closing timed out.")
+        self.assertFalse(TestCallbacks.scene.is_alive(), msg="Raytracing thread closing timed out.")
         TestCallbacks.is_alive = False
 
     @classmethod